move cast to geovista.common
