+++ src/geovista/__init__.py
@@ -28,7 +28,7 @@ from .cache import (  # noqa: F401
 from .common import vtk_warnings_off, vtk_warnings_on  # noqa: F401
 from .core import MeridianSlice, combine, cut_along_meridian  # noqa: F401
 from .crs import from_wkt, get_central_meridian, set_central_meridian  # noqa: F401
-from .filters import cast_UnstructuredGrid_to_PolyData, remesh  # noqa: F401
+from .filters import remesh  # noqa: F401
 from .geodesic import BBox, line, panel, wedge  # noqa: F401
 from .geometry import coastlines, load_coastlines  # noqa: F401
 from .geoplotter import GeoPlotter  # noqa: F401
